actions: move edit commit orchestration to core
Move EditBatch commit orchestration into core, apply planned object/link writes through ObjectStorage, and remove the old EditStorage port from core, SQLite, and Postgres.
