+++ packages/core/src/actions/errors.ts
@@ -1,3 +1,7 @@
 export class ActionDefinitionError extends Error {
   readonly name = "ActionDefinitionError"
 }
+
+export class ActionEditCommitError extends Error {
+  readonly name = "ActionEditCommitError"
+}
