+++ packages/core/src/index.ts
@@ -104,6 +104,7 @@ export type {
   ActionBuilder,
   ActionContext,
   ActionDefinition,
+  ActionEditCommitResult,
   ActionHandler,
   ActionParamConfig,
   ActionParamsBuilder,
@@ -114,6 +115,8 @@ export type {
   ActionTargetObject,
   ActionValidationContext,
   ActionValidator,
+  CommitActionEditBatchInput,
+  CommitActionEditBatchResult,
   GlobalActionContext,
   GlobalActionDefinition,
   GlobalActionHandler,
@@ -134,9 +137,11 @@ export type {
 } from "./actions"
 export {
   ActionDefinitionError,
+  ActionEditCommitError,
   ActionRegistry,
   ActionsRuntime,
   actionParam,
+  commitActionEditBatch,
   defineAction,
   isActionDefinition,
   isGlobalActionDefinition,
@@ -469,7 +474,6 @@ export type {
   AuthUserIdentityStore,
   AuthUserStore,
   CancelWorkflowInterventionInput,
-  CommitEditBatchInput,
   CompleteAuthSessionInput,
   CompleteMagicLinkSignInInput,
   CompleteOidcSignInInput,
@@ -483,8 +487,6 @@ export type {
   CreateOrUpdateAuthInvitationInput,
   CreateWorkflowInterventionInput,
   DefineMigrationsOptions,
-  EditCommitResult,
-  EditStorage,
   EnterActionRunPhaseInput,
   ExistsObjectsInput,
   ExistsObjectsResult,
@@ -637,11 +639,9 @@ export {
   canRequeueActionRunAfterEnqueueFailure,
   createTransactionStorageProxy,
   defineMigrations,
-  EditStorageError,
   finishActionRunPhase,
   InMemoryActionRunStorage,
   InMemoryAuthStorage,
-  InMemoryEditStorage,
   InMemoryObjectStorage,
   InMemoryPipelineRunStorage,
   InMemoryProjectionRunStorage,
@@ -658,6 +658,7 @@ export {
   isTerminalActionRun,
   migrateStorage,
   normalizeActionRunCommitDiff,
+  ObjectStorageError,
   PipelineRunError,
   ProjectionRunError,
   planMigrationSet,
