--- packages/core/src/storage/edits/errors.ts
@@ -1,6 +1,0 @@
-export class EditStorageError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = "EditStorageError"
-  }
-}
