--- packages/core/src/storage/edits/index.ts
@@ -1,3 +1,0 @@
-export { EditStorageError } from "./errors"
-export { InMemoryEditStorage } from "./in-memory"
-export type { CommitEditBatchInput, EditCommitResult, EditStorage } from "./types"
