+++ packages/core/src/storage/index.ts
@@ -81,9 +81,7 @@ export type {
   UserStatus,
 } from "./auth"
 export { AuthStorageError, InMemoryAuthStorage } from "./auth"
-export type { CommitEditBatchInput, EditCommitResult, EditStorage } from "./edits"
-export { EditStorageError, InMemoryEditStorage } from "./edits"
-export { ObjectNotFoundError, StorageTransactionError } from "./errors"
+export { ObjectNotFoundError, ObjectStorageError, StorageTransactionError } from "./errors"
 export type {
   DefineMigrationsOptions,
   MigrationCapableStorage,
@@ -254,7 +252,6 @@ export {
 import { AsyncLocalStorage } from "node:async_hooks"
 import { InMemoryActionRunStorage } from "./action-runs"
 import { InMemoryAuthStorage } from "./auth"
-import { InMemoryEditStorage } from "./edits"
 import { StorageTransactionError } from "./errors"
 import { InMemoryObjectStorage } from "./objects"
 import { InMemoryPipelineRunStorage } from "./pipeline-runs"
@@ -274,7 +271,6 @@ export class InMemoryStorage implements Storage {
   readonly timeseries = new InMemoryTimeseriesStorage()
   readonly auth = new InMemoryAuthStorage()
   readonly actionRuns = new InMemoryActionRunStorage()
-  readonly edits = new InMemoryEditStorage(this.objects, this.actionRuns)
   readonly syncRuns = new InMemorySyncRunStorage()
   readonly pipelineRuns = new InMemoryPipelineRunStorage()
   readonly projectionRuns = new InMemoryProjectionRunStorage()
