+++ packages/core/src/storage/types.ts
@@ -1,6 +1,5 @@
 import type { ActionRunStorage } from "./action-runs"
 import type { AuthStorage } from "./auth"
-import type { EditStorage } from "./edits"
 import type { ObjectStorage } from "./objects/types"
 import type { PipelineRunStorage } from "./pipeline-runs"
 import type { ProjectionRunStorage } from "./projection-runs"
@@ -66,9 +65,7 @@ export type {
   UserStatus,
 } from "./auth"
 export { AuthStorageError } from "./auth"
-export type { CommitEditBatchInput, EditCommitResult, EditStorage } from "./edits"
-export { EditStorageError } from "./edits"
-export { StorageTransactionError } from "./errors"
+export { ObjectStorageError, StorageTransactionError } from "./errors"
 export type {
   LinkDirection,
   ObjectLinkRow,
@@ -189,7 +186,6 @@ export interface StorageTransactionOptions {
 export interface Storage {
   objects: ObjectStorage
   timeseries: TimeseriesStorage
-  edits?: EditStorage
   auth?: AuthStorage
   actionRuns?: ActionRunStorage
   syncRuns?: SyncRunStorage
