+++ storage/sqlite/tests/sqlite-auth-storage.test.ts
@@ -86,7 +86,6 @@ function closeSqliteStorage(storage: SqliteStorage): void {
   storage.objects.close()
   storage.auth.close()
   storage.actionRuns.close()
-  storage.edits.close()
   storage.pipelineRuns.close()
   storage.projectionRuns.close()
   storage.workflowRuns.close()
