+++ storage/sqlite/tests/sqlite-pipeline-run-storage.test.ts
@@ -402,7 +402,6 @@ describe("SqlitePipelineRunStorage", () => {
       bundled.objects.close()
       bundled.auth.close()
       bundled.actionRuns.close()
-      bundled.edits.close()
       bundled.pipelineRuns.close()
       bundled.projectionRuns.close()
       bundled.workflowRuns.close()
