+++ storage/sqlite/tests/sqlite-storage-migrations.test.ts
@@ -446,7 +446,6 @@ function closeStorage(storage: SqliteStorage): void {
   storage.objects.close()
   storage.auth.close()
   storage.actionRuns.close()
-  storage.edits.close()
   storage.pipelineRuns.close()
   storage.projectionRuns.close()
   storage.workflowRuns.close()
