+++ storage/sqlite/tests/sqlite-workflow-intervention-storage.test.ts
@@ -182,7 +182,6 @@ function closeSqliteStorage(storage: SqliteStorage): void {
   storage.objects.close()
   storage.auth.close()
   storage.actionRuns.close()
-  storage.edits.close()
   storage.pipelineRuns.close()
   storage.projectionRuns.close()
   storage.workflowRuns.close()
