+++ storage/sqlite/tests/sqlite-workflow-run-storage.test.ts
@@ -481,7 +481,6 @@ function closeSqliteStorage(storage: SqliteStorage): void {
   storage.objects.close()
   storage.auth.close()
   storage.actionRuns.close()
-  storage.edits.close()
   storage.pipelineRuns.close()
   storage.projectionRuns.close()
   storage.workflowRuns.close()
